# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microservicio de autenticación para validar y crear JWT tokens",
        "title": "Bikes2Road Authentication API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@bikes2road.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/auth/v1",
    "paths": {
        "/health": {
            "get": {
                "description": "Verifica que el servicio esté funcionando",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Servicio funcionando correctamente",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http.HealthResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Autentica un usuario con email y password, retorna tokens JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login de usuario",
                "parameters": [
                    {
                        "description": "Credenciales de login",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bikes2road_authentication_internal_domain.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "schema": {
                            "$ref": "#/definitions/github_com_bikes2road_authentication_internal_domain.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Request inválido",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/oauth/login": {
            "post": {
                "description": "Autentica un usuario con Google OAuth y retorna tokens JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "OAuth login de usuario",
                "parameters": [
                    {
                        "description": "Credenciales de OAuth",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bikes2road_authentication_internal_ports.UserInfoOAuth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "schema": {
                            "$ref": "#/definitions/github_com_bikes2road_authentication_internal_domain.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Request inválido",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/refresh": {
            "post": {
                "description": "Genera un nuevo par de tokens usando un refresh token válido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refrescar token JWT",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bikes2road_authentication_internal_domain.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tokens refrescados",
                        "schema": {
                            "$ref": "#/definitions/github_com_bikes2road_authentication_internal_domain.RefreshResponse"
                        }
                    },
                    "400": {
                        "description": "Request inválido",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token inválido o expirado",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/validate": {
            "post": {
                "description": "Valida un token JWT y retorna sus claims si es válido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Validar token JWT",
                "parameters": [
                    {
                        "description": "Token a validar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bikes2road_authentication_internal_domain.ValidateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token validado",
                        "schema": {
                            "$ref": "#/definitions/github_com_bikes2road_authentication_internal_domain.ValidateResponse"
                        }
                    },
                    "400": {
                        "description": "Request inválido",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_bikes2road_authentication_internal_domain.JWTClaims": {
            "type": "object",
            "properties": {
                "aud": {
                    "description": "the `aud` (Audience) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.3",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "email": {
                    "type": "string"
                },
                "exp": {
                    "description": "the `exp` (Expiration Time) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.4",
                    "allOf": [
                        {
                            "$ref": "#/definitions/jwt.NumericDate"
                        }
                    ]
                },
                "iat": {
                    "description": "the `iat` (Issued At) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.6",
                    "allOf": [
                        {
                            "$ref": "#/definitions/jwt.NumericDate"
                        }
                    ]
                },
                "iss": {
                    "description": "the `iss` (Issuer) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.1",
                    "type": "string"
                },
                "jti": {
                    "description": "the `jti` (JWT ID) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.7",
                    "type": "string"
                },
                "nbf": {
                    "description": "the `nbf` (Not Before) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.5",
                    "allOf": [
                        {
                            "$ref": "#/definitions/jwt.NumericDate"
                        }
                    ]
                },
                "nick_name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "sub": {
                    "description": "the `sub` (Subject) claim. See https://datatracker.ietf.org/doc/html/rfc7519#section-4.1.2",
                    "type": "string"
                }
            }
        },
        "github_com_bikes2road_authentication_internal_domain.LoginRequest": {
            "type": "object",
            "required": [
                "email_or_nick_name",
                "password"
            ],
            "properties": {
                "email_or_nick_name": {
                    "type": "string",
                    "example": "[EMAIL_ADDRESS] | johndoe"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "T3st123@"
                }
            }
        },
        "github_com_bikes2road_authentication_internal_domain.LoginResponse": {
            "type": "object",
            "properties": {
                "tokens": {
                    "$ref": "#/definitions/github_com_bikes2road_authentication_internal_domain.TokenPair"
                },
                "user": {
                    "$ref": "#/definitions/github_com_bikes2road_authentication_internal_domain.UserInfo"
                }
            }
        },
        "github_com_bikes2road_authentication_internal_domain.RefreshRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "github_com_bikes2road_authentication_internal_domain.RefreshResponse": {
            "type": "object",
            "properties": {
                "tokens": {
                    "$ref": "#/definitions/github_com_bikes2road_authentication_internal_domain.TokenPair"
                }
            }
        },
        "github_com_bikes2road_authentication_internal_domain.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "description": "Segundos hasta expiración",
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
        "github_com_bikes2road_authentication_internal_domain.UserInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "github_com_bikes2road_authentication_internal_domain.ValidateRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "github_com_bikes2road_authentication_internal_domain.ValidateResponse": {
            "type": "object",
            "properties": {
                "claims": {
                    "$ref": "#/definitions/github_com_bikes2road_authentication_internal_domain.JWTClaims"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "github_com_bikes2road_authentication_internal_ports.UserInfoOAuth": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "internal_adapters_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_adapters_http.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "jwt.NumericDate": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}